# src/docs/swagger.yaml
definitions:
...
  models.Meeting:
    properties:
      actual_end_time:
        type: string
      actual_start_time:
        type: string
      all_participants_responded:
        type: boolean
      created_at:
        type: string
      description:
        type: string
      end_time:
        type: string
      hour:
        type: integer
      id:
        type: integer
      is_confirmed:
        type: boolean
      is_online:
        type: boolean
      is_onsite:
        type: boolean
      place:
        type: string
      start_time:
        type: string
      title:
        type: string
      updated_at:
        type: string
      url:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/restricted/meetings/{id}:
    get:
      description: Get meetings by meeting id
      parameters:
      - description: Meeting ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Meeting'
        "400":
          description: Error message
          schema:
            type: string
      summary: Get meetings by meeting id
      tags:
      - meetings
  /api/restricted/meetings/user/{user_id}:
    get:
      description: Get meetings by user id
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Meeting'
        "400":
          description: Error message
          schema:
            type: string
      summary: Get meetings by user id
      tags:
      - meetings
...
  /api/signup:
    post:
      consumes:
      - application/json
      description: Create a new meeting
      parameters:
      - description: Meeting details
        in: body
        name: meeting
        required: true
        schema:
          $ref: '#/definitions/models.Meeting'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Meeting'
        "400":
          description: Error message
          schema:
            type: string
      summary: Create a new meeting
      tags:
      - meetings
  /meetings/{id}:
    put:
      description: Update meeting by meeting id
      parameters:
      - description: Meeting ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Meeting'
        "400":
          description: Error message
          schema:
            type: string
      summary: Update meeting by meeting id
      tags:
      - meetings
  /meetings/guest/confirmed/{user_id}:
    get:
      description: Get meetings that are not responded by host by user id
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Meeting'
        "400":
          description: Error message
          schema:
            type: string
      summary: Get meetings that are confirmed by guest by user id
      tags:
      - meetings
  /meetings/guest/not-confirmed/{user_id}:
    get:
      description: Get meetings that are not confirmed by guest by user id
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Meeting'
        "400":
          description: Error message
          schema:
            type: string
      summary: Get meetings that are not confirmed by guest by user id
      tags:
      - meetings
  /meetings/guest/not-responded/{user_id}:
    get:
      description: Get meetings that are not responded by guest by user id
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Meeting'
        "400":
          description: Error message
          schema:
            type: string
      summary: Get meetings that are not responded by guest by user id
      tags:
      - meetings
  /meetings/host/confirmed/{user_id}:
    get:
      description: Get meetings that are confirmed for host by user id
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Meeting'
        "400":
          description: Error message
          schema:
            type: string
      summary: Get meetings that are confirmed for host by user id
      tags:
      - meetings
  /meetings/host/not-confirmed/{user_id}:
    get:
      description: Get meetings that are not confirmed for host by user id
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Meeting'
        "400":
          description: Error message
          schema:
            type: string
      summary: Get meetings that are not confirmed for host by user id
      tags:
      - meetings
  /meetings/host/not-responded/{user_id}:
    get:
      description: Get meetings that are not responded by host by user id
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Meeting'
        "400":
          description: Error message
          schema:
            type: string
      summary: Get meetings that are not responded by host by user id
      tags:
      - meetings
...
swagger: "2.0"